Fixed naming convention for BlogList component - was no longer a helper function since it returns JSX, and added a window.confirm to delete blog posts

+++ part5/bloglist-frontend/src/components/Blog.js
@@ -45,6 +45,9 @@ const Blog = ({ blog, blogs, setBlogs, user, setSuccessMessage }) => {
     }
 
     const handleDelete = async blog => {
+      if (!window.confirm(`Do you really want to delete ${blog.title}?`)) {
+        return
+      }
       const { id, title } = blog
       const removedBlog = await blogService.remove(id)
       if (removedBlog === 204) {
